Changes for 2015-04-11 release
New features
- Moved to ControlClick method, from MouseClick method, so that bot does
not interfere with mouse, and normal use of the computer can continue.
The bot "transparently" operates in the BlueStacks window now.
BlueStacks still needs to be fully visible and not obscured, as the bot
needs to read the pixels on the window to decide what to do.  Thanks,
ClashGameBot! (http://gamebot.org/)

Fixed bugs
- Fixed bug in detecting amount of dark elixir won at end of battle.
- Adjusted logging so that all debug statements are directed to a file.
- Adjusted detection of live replay screen to allow slightly more error
in color detection.
- Fixed rarely occurring timeout error when processing the end battle
summary screen.