Correcting error in storage calculation logic
Left off one of the situations in the if-then-else tree when calculating
loot penalty.  Forgot to include the situation where we were attacking a
town hall equal or higher to us.  I have typically always attacked with
the match filter to set to 1 less than my TH level, so never noticed
this before.